move generic functions to ecto_sparkles
