--- lib/repo/changesets.ex
@@ -1,9 +1,0 @@
-defmodule Bonfire.Common.Changesets do
-
-  alias Ecto.Changeset
-
-  def error(changeset, []), do: changeset
-  def error(changeset, [{k, v} | errors]),
-    do: error(Changeset.add_error(changeset, k, v), errors)
-
-end
